+++ src/PyCRTools/modules/core/utils.py
@@ -171,3 +171,58 @@ def readParfiles(parfile):
         else:
             print "Error: parameter files",parfile,"not found."
     return pardict
+
+## {{{ http://code.activestate.com/recipes/208993/ (r1)
+#!/usr/bin/env python
+#
+# Author: Cimarron Taylor
+# Date: July 6, 2003
+# File Name: relpath.py
+# Program Description: Print relative path from /a/b/c/d to /a/b/c1/d1
+
+#
+# helper functions for relative paths
+#
+import os
+
+# def pathsplit(p, rest=[]):
+#     (h,t) = os.path.split(p)
+#     if len(h) < 1: return [t]+rest
+#     if len(t) < 1: return [h]+rest
+#     return pathsplit(h,[t]+rest)
+
+def pathsplit(path):
+    """ This version, in contrast to the original version, permits trailing
+    slashes in the pathname (in the event that it is a directory).
+    It also uses no recursion """
+    return path.split(os.path.sep)
+
+def commonpath(l1, l2, common=[]):
+    if len(l1) < 1: return (common, l1, l2)
+    if len(l2) < 1: return (common, l1, l2)
+    if l1[0] != l2[0]: return (common, l1, l2)
+    return commonpath(l1[1:], l2[1:], common+[l1[0]])
+
+def relpath(p1, p2):
+    #First deal with the issue if both are relative paths
+    if len(p1)==0 or not p1[0]==os.path.sep:
+        if len(p2)==0 or not p2[0]==os.path.sep:
+            if p2[:len(p1)]==p1:
+                s=p2[len(p1):]
+                if len(s)>0 and s[0]==os.path.sep: s=s[1:]
+                return s
+            
+    (common,l1,l2) = commonpath(pathsplit(p1), pathsplit(p2))
+    p = []
+    if len(l1) > 0:
+        p = [ ('..'+os.path.sep) * len(l1) ]
+    p = p + l2
+    return os.path.join( *p )
+
+
+#def test(p1,p2):
+#    print "from", p1, "to", p2, " -> ", relpath(p1, p2)
+#
+#if __name__ == '__main__':
+#    test('/a/b/c/d', '/a/b/c1/d1')
+## end of http://code.activestate.com/recipes/208993/ }}}